backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Supabase
supabase==2.3.4

# LangChain and AI
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.17
openai==1.10.0

# Document processing
pypdf==4.0.1
python-pptx==0.6.23
python-docx==1.1.0
unstructured==0.12.2

# Validation
pylint==3.0.3
black==24.1.1
isort==5.13.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiohttp==3.9.3
tenacity==8.2.3
tiktoken==0.5.2

# Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
